@@ -34,7 +38,7 @@
         "url": "https://github.com/react-in-action/letters-social/issues"
     },
     "engines": {
-        "node": "8.1.4",
+        "node": "8.4.0",
         "npm": "5.3.0"
     },
     "jest": {
@@ -83,11 +88,11 @@
         "moment": "2.18.1",
         "morgan": "1.8.2",
         "node-uuid": "1.4.8",
-        "npm-run-all": "4.0.2",
+        "npm-run-all": "4.1.1",
         "ora": "1.3.0",
-        "react": "15.6.1",
-        "react-dom": "15.6.1",
-        "react-overlays": "0.8.0",
+        "parse-link-header": "1.0.1",
+        "react": "16.0.0-beta.5",
+        "react-dom": "16.0.0-beta.5",
         "react-redux": "5.0.6",
         "react-router": "3.0.5",
         "react-router-redux": "3.0.0",
