@@ -1,20 +1,24 @@
-const { join } = require('path');
-const { name, internet, lorem, date, random } = require('faker');
-const { promisify } = require('bluebird');
-const mkdirp = require('mkdirp');
-const { sample, sampleSize, random: rand } = require('lodash');
-const { v4: uuid } = require('node-uuid');
-const { writeFile } = require('fs');
-const ora = require('ora');
+import { join } from 'path';
+import fetch from 'isomorphic-fetch';
+import { name, internet, lorem, date, random } from 'faker';
+import starwars from 'starwars';
+import { promisify } from 'bluebird';
+import mkdirp from 'mkdirp';
+import _ from 'lodash';
+import uuid from 'uuid/v4';
+import { writeFile } from 'fs';
+import ora from 'ora';
 
-const { categories } = require('./constants');
+const { sample, sampleSize, random: rand } = _;
+
+const { categories, swapiURL } = require('./constants');
 
 const write = promisify(writeFile);
 
 function generateProfilePicture() {
     const pics = [];
     for (let i = 0; i < 15; i++) {
-        pics.push(`https://drtzvj8zd0k9x.cloudfront.net/assets/profile-pictures/${i + 1}.png`);
+        pics.push(`https://react-sh.s3.amazonaws.com/assets/profile-pictures/${i + 1}.png`);
     }
     return function selectRandomProfilePicture() {
         return sample(pics);
@@ -33,29 +37,27 @@ function generateShareablePicture() {
 }
 const createShareableImage = generateShareablePicture();
 
-module.exports.User = class User {
-    constructor() {
-        this.email = internet.email();
-        this.firstName = name.firstName();
-        this.id = uuid();
-        this.profilePicture = createProfilePicture();
-        this.lastName = name.lastName();
-        this.password = internet.password();
-        this.username = internet.userName();
+module.exports.UserFactory = {
+    generate: async () => {
+        const newUser = {};
+        const countRes = await fetch(`${swapiURL}/people`);
+        const { count } = await countRes.json();
+        const personRes = await fetch(`${swapiURL}/people/${rand(0, count)}`);
+        const swapiPerson = await personRes.json();
+        newUser.name = swapiPerson.name || name.findName();
+        newUser.email = internet.email();
+        newUser.id = uuid();
+        newUser.profilePicture = createProfilePicture();
+        return newUser;
     }
 };
 
 // possible categories
-
-function returnCategories() {
-    return sampleSize(categories, rand(1, 15));
-}
 module.exports.Post = class Post {
-    constructor(user) {
+    constructor(userId) {
         this.id = uuid();
-        this.categories = returnCategories();
-        this.comments = [];
-        this.content = lorem.paragraph(sample([1, 2, 3]));
+        this.categories = sampleSize(categories, rand(1, 15));
+        this.content = starwars();
         this.date = date.recent(sample([1, 2, 3, 4, 5, 10, 15]));
         this.image = Math.random() * 10 > 3 ? null : createShareableImage();
         this.likes = random.number(1, 100);
@@ -66,56 +68,54 @@ module.exports.Post = class Post {
                   title: lorem.words(rand(1, 5)),
                   description: lorem.sentences(rand(1, 2), '. ')
               };
-        this.user = user;
+        this.userId = userId;
     }
 };
 
 module.exports.Comment = class Comment {
-    constructor(user) {
+    constructor(userId, postId) {
         this.id = uuid();
         this.content = lorem.paragraph(sample([1, 2, 3]));
         this.date = date.recent(sample([1, 2, 3, 4, 5, 10, 15]));
         this.likes = random.number(1, 100);
-        this.user = user;
+        this.postId = postId;
+        this.userId = userId;
     }
 };
 
-function generateUsers(n) {
-    const users = [];
-    for (let i = 0; i < n; i++) {
-        users.push(new module.exports.User());
-    }
-    return users;
+async function generateUsers(n) {
+    return await Promise.all(
+        [...Array(n).keys()].map(async () => await module.exports.UserFactory.generate())
+    );
 }
 
-function generateComments(n, users) {
+function generateComments(n, users, posts) {
     const comments = [];
     for (let i = 0; i < n; i++) {
-        comments.push(new module.exports.Comment(sample(users)));
+        const userIds = users.map(u => u.id);
+        const postIds = posts.map(u => u.id);
+        const randomUser = sample(userIds);
+        const randomPost = sample(postIds);
+        comments.push(new module.exports.Comment(randomUser, randomPost));
     }
     return comments;
 }
 
-function generatePosts(n, users, comments) {
+function generatePosts(n, users) {
     const posts = [];
     for (let i = 0; i < n; i++) {
         const newPost = new module.exports.Post(sample(users));
-        newPost.comments = sampleSize(comments, random.number(7));
         posts.push(newPost);
     }
     return posts;
 }
 
-module.exports.seed = function seed(
-    nUsers = 500,
-    nPosts = 2000,
-    nComments = 750
-) {
+(async function seed(nUsers = 10, nPosts = 20, nComments = 25) {
     const spinner = ora('Generating sample data...').start();
     mkdirp.sync(join(__dirname, 'seed'));
-    const users = generateUsers(nUsers);
-    const comments = generateComments(nComments, users);
-    const posts = generatePosts(nPosts, users, comments);
+    const users = await generateUsers(nUsers);
+    const posts = generatePosts(nPosts, users);
+    const comments = generateComments(nComments, users, posts);
     Promise.all([
         write(
             join(__dirname, 'seed', 'db.json'),
@@ -134,6 +134,4 @@ module.exports.seed = function seed(
             console.error(err);
             setTimeout(() => spinner.stop());
         });
-};
-
-module.exports.seed();
+})();
