@@ -13,9 +13,7 @@ const Logo = props => {
                 <span style={{ color: '#417505' }}>e</span>
                 <span style={{ color: '#D0021B' }}>r</span>
                 <span style={{ color: '#F8E71C' }}>s</span>
-                {props.logoOnly
-                    ? null
-                    : <small className="social">.social</small>}
+                {props.logoOnly ? null : <small className="social">.social</small>}
             </div>
         </Link>
     );
