@@ -1,61 +1,48 @@
-import React, { Component } from 'react';
-
-import { logout, firebase } from '../../backend';
+import React from 'react';
+import { connect } from 'react-redux';
+import { logout } from '../../actions/auth';
 import Link from '../router/Link';
 import Logo from './Logo';
 
-class Navigation extends Component {
-    constructor(props) {
-        super(props);
-        this.handleLogout = this.handleLogout.bind(this);
-        this.state = {
-            user: null
-        };
-    }
-    componentDidMount() {
-        firebase.auth().onAuthStateChanged(user => this.setState({ user }));
-    }
-
-    handleLogout() {
-        logout().then(() => this.setState({ user: null }));
-    }
-
-    render() {
-        return (
-            <nav className="navbar navbar-default navbar-fixed-top">
-                <div className="container">
-                    <div className="row middle-xs center-xs around-xs">
-                        <div className="col-sm-4 col-sm-offset-4 col-xs-8">
-                            <Logo logoOnly={false} />
-                        </div>
-                        <div className="col-sm-4 col-xs-4">
-                            {this.state.user
-                                ? <div>
-                                      <div className="btn-group" role="group">
-                                          <button
-                                              onClick={this.handleLogout}
-                                              className="btn btn-default"
-                                          >
-                                              <i className="fa fa-sign-out" />
-                                          </button>
-                                      </div>
-                                  </div>
-                                : <div>
-                                      <Link to="/login">
-                                          <button
-                                              type="button"
-                                              className="btn btn-default"
-                                          >
-                                              Log in or sign up
-                                          </button>
-                                      </Link>
-                                  </div>}
-                        </div>
-                    </div>
+const Navigation = ({ user, handleLogout }) => (
+    <nav className="navbar navbar-default navbar-fixed-top">
+        <div className="container">
+            <div className="row middle-xs center-xs around-xs">
+                <div className="col-sm-4 col-sm-offset-4 col-xs-8">
+                    <Logo logoOnly={false} />
                 </div>
-            </nav>
-        );
-    }
-}
+                <div className="col-sm-4 col-xs-4">
+                    {user.authenticated ? (
+                        <span className="user-nav-widget">
+                            <span>{user.name}</span>
+                            <img
+                                width={40}
+                                className="img-circle"
+                                src={user.avatar}
+                                alt={user.name}
+                            />
+                            <i onClick={handleLogout} className="fa fa-sign-out" />
+                        </span>
+                    ) : (
+                        <div>
+                            <Link to="/login">
+                                <button type="button" className="btn btn-default">
+                                    Log in or sign up
+                                </button>
+                            </Link>
+                        </div>
+                    )}
+                </div>
+            </div>
+        </div>
+    </nav>
+);
 
-export default Navigation;
+export default connect(
+    state => ({ user: state.user }),
+    dispatch => ({
+        handleLogout() {
+            dispatch(logout());
+        }
+    })
+)(Navigation);
