@@ -1,15 +1,19 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import { connect } from 'react-redux';
 
 import Nav from '../components/nav/Navbar';
 
+/**
+ * The app component serves as a root for the project and renders either children or the error state
+ * @method App
+ * @module letters/components
+ */
 const App = props => {
     return (
         <div className="app">
             <Nav />
-            <div className="container-fluid">
-                {props.children}
-            </div>
+            {props.error || props.children}
         </div>
     );
 };
@@ -18,4 +22,8 @@ App.propTypes = {
     children: PropTypes.node
 };
 
-export default App;
+export default connect(state => {
+    return {
+        error: state.error
+    };
+})(App);
