@@ -1,10 +1,9 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { AutoAffix } from 'react-overlays';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import { getPost, getPosts } from '../actions/posts';
+import { getPostByID, getPostsForPage } from '../actions/posts';
 
 import Ad from '../components/ad/Ad';
 import Post from '../components/post/Post';
@@ -20,34 +19,23 @@ class SinglePost extends Component {
     componentDidMount() {
         // If there's no posts already loaded, load them
         if (!this.props.post) {
-            this.props.actions.getPost(this.props.router.params.post);
+            this.props.actions.getPostByID(this.props.router.params.post);
         }
     }
 
     render() {
         return (
-            <div className="container-fluid">
-                <div className="row">
-                    <div className="col-sm-3">
-                        <Link to="/">
-                            <div className="pull-right">
-                                <br />
-                                <i className="fa fa-arrow-left" /> Back
-                            </div>
-                        </Link>
-                    </div>
-                    <div className="col-xs-12 col-sm-6">
-                        <Post forceOpen post={this.props.post} />
-                    </div>
-                    <div className="col-sm-2 col-xs-12 last-xs">
-                        <AutoAffix viewportOffsetTop={50} container={this}>
-                            <Ad
-                                url="https://www.manning.com/books/react-in-action"
-                                imageUrl="https://drtzvj8zd0k9x.cloudfront.net/assets/ads/react+in+action+meap+ad.png"
-                            />
-                        </AutoAffix>
-                    </div>
-                </div>
+            <div className="single-post">
+                <Link to="/">
+                    <span>
+                        <i className="fa fa-arrow-left" /> Back
+                    </span>
+                </Link>
+                <Post forceOpen post={this.props.post} />
+                <Ad
+                    url="https://www.manning.com/books/react-in-action"
+                    imageUrl="https://drtzvj8zd0k9x.cloudfront.net/assets/ads/react+in+action+meap+ad.png"
+                />
             </div>
         );
     }
@@ -64,7 +52,7 @@ export default connect(
     },
     dispatch => {
         return {
-            actions: bindActionCreators({ getPost, getPosts }, dispatch)
+            actions: bindActionCreators({ getPostByID, getPostsForPage }, dispatch)
         };
     }
 )(SinglePost);
