@@ -2,11 +2,17 @@ import { combineReducers } from 'redux';
 import { loading } from './loading';
 import { posts, postIds } from './posts';
 import { user } from './user';
+import { pagination } from './pagination';
 
+/**
+ * Root reducer for project
+ * @module letters/reducers
+ */
 const rootReducer = combineReducers({
     loading,
-    posts,
+    pagination,
     postIds,
+    posts,
     user
 });
 
