@@ -1,5 +1,12 @@
 import fetch from 'isomorphic-fetch';
 
+/**
+ * Creates a post with the given payload
+ * @method createPost
+ * @module letters/shared/http
+ * @param  {object}   payload Post payload
+ * @return {Response}           Fetch Response
+ */
 export function createPost(payload) {
     if (!payload) {
         throw new Error('You must provide a payload when creating a new post');
@@ -14,17 +21,27 @@ export function createPost(payload) {
     };
 
     // Send the new post to the API
-    return fetch(`${process.env.ENDPOINT}/posts`, requestOptions).then(res =>
-        res.json()
-    );
+    return fetch(`${process.env.ENDPOINT}/posts`, requestOptions);
 }
 
-export function fetchPosts(n) {
-    return fetch(
-        `${process.env.ENDPOINT}/posts?_limit=${n}&_sort=date&_order=DESC`
-    ).then(res => res.json());
+/**
+ * Fetch posts from the API
+ * @module letters/shared/http
+ * @method fetchPosts
+ * @param  {string}   endpoint URL provided by Redux; the API will yield further endpoints we can access via the Link Header (https://www.w3.org/wiki/LinkHeader)
+ * @return {Response}          Fetch API Response
+ */
+export function fetchPosts(endpoint) {
+    return fetch(endpoint);
 }
 
+/**
+ * Fetch a post from the API
+ * @module letters/shared/http
+ * @method fetchPost
+ * @param  {string}  id post ID
+ * @return {Response}     Fetch Response object
+ */
 export function fetchPost(id) {
-    return fetch(`${process.env.ENDPOINT}/posts/${id}`).then(res => res.json());
+    return fetch(`${process.env.ENDPOINT}/posts/${id}`);
 }
